Move Dune query-related function to Dune source module
